+++ test_data/test_parser/includes/sample_code.js
@@ -1,0 +1,5 @@
+var str = "sample code";
+var i = 0;
+for (i = 0; i < 10; i++) {
+  str += i;
+}