+++ test_data/test_parser/includes/sample_code.py
@@ -1,0 +1,7 @@
+    # start example 1
+    print("test dedent")
+    # end example 1
+
+    # start example 2
+    print("hello world")
+    # end example 2