+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for literal includes (DOP-876).
+
 ## [v0.4.8] - 2020-05-27
 
 ### Added